+++ crates/client/rpc/src/methods/read/get_block_transaction_count.rs
@@ -8,8 +8,8 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use starknet_core::types::BlockId;
 
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
 use crate::Starknet;
 
 /// Get the Number of Transactions in a Given Block