+++ crates/client/rpc/src/methods/read/get_block_with_receipts.rs
@@ -14,9 +14,10 @@ use starknet_core::types::{
     TransactionWithReceipt,
 };
 
+use super::get_transaction_receipt::get_transaction_receipt_finalized;
+use crate::deoxys_backend_client::get_block_by_block_hash;
 use super::get_transaction_receipt::receipt;
 use crate::errors::StarknetRpcApiError;
-use crate::madara_backend_client::get_block_by_block_hash;
 use crate::utils::block::{
     l1_da_mode, l1_data_gas_price, l1_gas_price, new_root, parent_hash, sequencer_address, starknet_version, timestamp,
 };