+++ crates/node/src/configs.rs
@@ -1,3 +1,6 @@
+use std::path::PathBuf;
+
+use sc_service::Configuration;
 use serde::Deserialize;
 
 #[derive(Deserialize)]
@@ -11,4 +14,9 @@ pub struct FileInfos {
     pub name: String,
     pub sha3_256: Option<String>,
     pub url: Option<String>,
+}
+
+/// Returns the path to the database of the node.
+pub fn db_config_dir(config: &Configuration) -> PathBuf {
+    config.base_path.config_dir(config.chain_spec.id())
 }